Return $null if array is an index
According to the docs, array fields are not supported for indexes. II had an error by adding an array as an index and then it's very hard to recover from it (`field.removeIndex(...)` will not work because the index is already stored in storage.

The easiest solution I can see is to return $null from the `projectForUnique` function when the indexed field is an array. At least, it worked for me. Cheers